requirements.txt

beautifulsoup4==4.9.3
requests==2.25.1
PyYAML
pdfkit==0.6.1
